Test jest (#1)
* utilizando jest nos produtos

* continucao de teste com jest

* continucao2 de teste com jest

* config jest em todo codigo

+++ src/controllers/categoryController.js
@@ -9,26 +9,33 @@ module.exports = {
         const { name } = req.body;
 
         try {
+            if (!name) {
+                throw { type: 'AccessDenied', status: 401, message: "invalid Name" };
+            }
 
             let category = await categoryModel.findOne({ name })
 
             if (category) {
-                return res.status(401).json({
-                    message: "Catagory alread axists",
-                    Categoria: {
-                        id: category._id,
-                        name: category.name
+                throw {
+                    type: 'AccessDenied', status: 401, message: {
+                        message: "Catagory alread axists",
+                        Categoria: {
+                            id: category._id,
+                            name: category.name
+                        }
                     }
-                })
+                };
             }
             category = await categoryModel.create({ name });
 
-            return res.status(201).json({ category })
+            return res.status(201).json(category)
         } catch (error) {
-            return res.status(500).json({ message: error.message })
+
+            return res.status(error.status || 500).json({ message: error.message })
         }
 
     },
+
     async filter(req, res) {
         const { name } = req.query;
         try {
@@ -38,44 +45,56 @@ module.exports = {
          * option "x" ignora os espacos
          * option "i" ignora difença entre maiusculo e minusculo
           */
+            if (!name) {
+                throw { type: 'AccessDenied', status: 404, message: 'Invalid name' };
+            }
+
             const category = await categoryModel.find({ name: { $regex: name, $options: "i" } }).
                 limit(5).select(['name', 'id', 'products']);
 
             if (category.length < 1) {
-                return res.status(404).json({ message: "Category does not exists" })
+                throw { type: 'AccessDenied', status: 404, message: 'Category does not exists' };
             }
 
             return res.status(200).json(category)
         } catch (error) {
 
-            return res.status(500).json({ message: error.message })
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
+
     async update(req, res) {
         const { idCategory } = req.params;
         const { name } = req.body;
 
         try {
             const isValidId = mongoose.Types.ObjectId.isValid(idCategory);
+
             if (!isValidId) {
-                return res.status(401).json({ message: "Category does not exist" });
+                throw { type: 'AccessDenied', status: 404, message: 'Category does not exist' };
+            }
+            if (!name) {
+                throw { type: 'AccessDenied', status: 401, message: 'Invalid Name' };
             }
 
             const category = await categoryModel.findOne({ _id: idCategory });
 
             if (!category) {
-                return res.status(401).json({ message: "Category does not exist" })
+                throw { type: 'AccessDenied', status: 404, message: 'Category does not exist' };
             }
 
             const nameCategoryAlreadyExists = await categoryModel.findOne({ name })
 
-            if (nameCategoryAlreadyExists._id == category._id) {
-                return res.status(401).json({
-                    message: "Category already registered", Categoria: {
-                        id: nameCategoryAlreadyExists._id,
-                        name: nameCategoryAlreadyExists.name
+            if (nameCategoryAlreadyExists && nameCategoryAlreadyExists._id != category._id) {
+                throw {
+                    type: 'AccessDenied', status: 404,
+                    message: {
+                        message: "Category already registered", Categoria: {
+                            id: nameCategoryAlreadyExists._id,
+                            name: nameCategoryAlreadyExists.name
+                        }
                     }
-                })
+                }
             }
 
             category.name = name;
@@ -84,10 +103,10 @@ module.exports = {
 
             return res.status(201).json({ message: `Category ${category.name} edited with sucess` })
         } catch (error) {
-
-            return res.status(500).json({ message: error.message })
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
+
     async index(req, res) {
         try {
             const categoryList = await categoryModel.find().select(['name', 'products']);
@@ -98,22 +117,30 @@ module.exports = {
             return res.status(500).json({ message: error.message });
         }
     },
+
     async destroy(req, res) {
         const { idCategory } = req.params;
 
         try {
+            const isValidId = mongoose.Types.ObjectId.isValid(idCategory);
+
+            if (!isValidId) {
+                throw { type: 'AccessDenied', status: 401, message: 'Category does not exist' };
+            }
+
             const category = await categoryModel.findOne({ _id: idCategory });
 
             if (!category) {
-                return res.status(401).json({ message: "Category does not exist" })
+                throw { type: 'AccessDenied', status: 404, message: 'Category does not exist' };
             }
 
             await categoryModel.deleteMany({ _id: idCategory }).populate('products');
 
-            return res.status(200).json({ message: `Category Delete with sucess` })
+            return res.status(203).send();
+
         } catch (error) {
 
-            return res.status(400).json({ message: error.message })
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
 }
