+++ src/controllers/productController.js
@@ -1,12 +1,13 @@
 const productModel = require('../models/product');
 const categoryModel = require('../models/category');
 const yup = require('yup');
+const mongoose = require('mongoose');
 
 
 const schema = yup.object().shape({
     title: yup.string().required().min(2),
     description: yup.string(),
-    price: yup.number().positive(),
+    price: yup.number().positive().required(),
     category: yup.string().required().min(2)
 });
 
@@ -14,14 +15,13 @@ module.exports = {
 
     async create(req, res) {
         const { title, description, price, category } = req.body;
-
         try {
             await schema.validate({ title, description, price, category });
 
             const productAlreadExists = await productModel.findOne({ title });
 
             if (productAlreadExists) {
-                return res.status(401).json({ message: "Product alread Exists" })
+                throw { type: 'AccessDenied', status: 401, message: 'Product alread Exists' };
             }
 
             let categoryObject = await categoryModel.findOne({ name: category })
@@ -41,34 +41,41 @@ module.exports = {
 
             await categoryObject.save();
 
-            return res.status(201).json({ productRegistred });
+            return res.status(201).json(productRegistred);
         } catch (error) {
             if (error instanceof yup.ValidationError) {
-                res.status(400).json({ message: error.message });
+                return res.status(401).json({ message: error.message });
             }
-            return res.status(500).json({ message: error.message });
+            return res.status(error.status || 500).json({ message: error.message });
         }
     },
+
     async index(req, res) {
 
         try {
             const productList = await productModel.find().populate('assingTo', 'name');;
 
-            return res.status(200).json({ productList })
+            return res.status(200).json(productList)
         } catch (error) {
 
-            res.status(400).json({ message: error.message })
+            res.status(500).json({ message: error.message })
         }
     },
+
     async update(req, res) {
         const { idProduct } = req.params;
         const { title, description, price, category } = req.body;
 
         try {
+            const isvalid = await mongoose.isValidObjectId(idProduct);
+            if (!isvalid) {
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
+            }
+
             const productIdParams = await productModel.findById(idProduct);
 
             if (!productIdParams) {
-                return res.status(404).json({ message: "Product does not exists" })
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
             }
 
             /***
@@ -78,7 +85,7 @@ module.exports = {
              */
             const titleAlreadExist = await productModel.findOne({ title })
             if (titleAlreadExist && (titleAlreadExist._id.toString() !== productIdParams._id.toString())) {
-                return res.status(404).json({ message: "Product title alread exists", titleAlreadExist })
+                throw { type: 'AccessDenied', status: 401, message: { message: "Product title alread exists", titleAlreadExist } };
             }
 
             /**
@@ -110,12 +117,12 @@ module.exports = {
             await categoryObject.save();
 
             return res.status(201).json({ product })
-        } catch (error) {
 
-            console.log(error);
-            return res.status(400).json({ message: error.message })
+        } catch (error) {
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
+
     async filter(req, res) {
         const { title } = req.query;
         /**
@@ -124,8 +131,11 @@ module.exports = {
          * option "x" ignora os espacos
          * option "i" ignora difença entre maiusculo e minusculo
           */
-
         try {
+            if (!title) {
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
+            }
+
             const product = await productModel.find({ title: { $regex: title, $options: "i" } }).
                 limit(5).select(["_id",
                     "title",
@@ -134,31 +144,37 @@ module.exports = {
                 ]).populate('assingTo', 'name');
 
             if (product.length < 1) {
-                return res.status(404).json({ message: "Product does not exists" })
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
             }
 
-            return res.status(200).json({ product })
+            return res.status(200).json(product)
         } catch (error) {
 
-            return res.status(500).json({ message: error.message })
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
+
     async destroy(req, res) {
         const { idProduct } = req.params;
 
         try {
+            const isvalid = await mongoose.isValidObjectId(idProduct);
+            if (!isvalid) {
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
+            }
+
             const product = await productModel.findOne({ _id: idProduct });
 
             if (!product) {
-                return res.status(401).json({ message: "Product does not exist" })
+                throw { type: 'AccessDenied', status: 404, message: 'Product does not exists' };
             }
 
             await productModel.deleteMany({ _id: idProduct }).populate('assyngTo');
 
-            return res.status(200).json({ message: `Product Delete with sucess` })
+            return res.status(203).json({ message: `Product Delete with sucess` })
         } catch (error) {
 
-            return res.status(400).json({ message: error.message })
+            return res.status(error.status || 500).json({ message: error.message })
         }
     },
 }
